Move stencils to a dedicated directory
