+++ packages/core/src/view/geometry/stencil/StencilShape.ts
