+++ packages/core/src/view/geometry/stencil/StencilShapeRegistry.ts
